Send messages with all attributes for config files
