+++ src/headers/syscheck_op.h
@@ -13,6 +13,21 @@
 
 #include "analysisd/eventinfo.h"
 
+#ifndef WIN32
+
+#include <sys/types.h>
+#include <pwd.h>
+#include <grp.h>
+
+#else
+
+#include "shared.h"
+#include "aclapi.h"
+
+#define BUFFER_LEN 1024
+
+#endif
+
 /* Fields for rules */
 #define SK_FILE    0
 #define SK_SIZE    1
@@ -86,5 +101,18 @@ int sk_decode_sum(sk_sum_t *sum, char *c_sum);
 void sk_fill_event(Eventinfo *lf, const char *f_name, const sk_sum_t *sum);
 
 int sk_build_sum(const sk_sum_t * sum, char * output, size_t size);
+
+
+#ifndef WIN32
+
+const char* get_user(__attribute__((unused)) const char *path, int uid);
+const char* get_group(int gid);
+
+#else
+
+const char *get_user(const char *path, __attribute__((unused)) int uid);
+const char *get_group(__attribute__((unused)) int gid);
+
+#endif
 
 #endif
