+++ src/syscheckd/syscheck.h
@@ -63,7 +63,4 @@ int c_read_file(const char *file_name, const char *oldsum, char *newsum) __attri
 int send_syscheck_msg(const char *msg) __attribute__((nonnull));
 int send_rootcheck_msg(const char *msg) __attribute__((nonnull));
 
-const char* get_user(const char *path, int uid);
-const char* get_group(int gid);
-
 #endif
